fix(serve-sim): support Xcode 27 (SimulatorKit relocation)
Xcode 27 moved SimulatorKit.framework from Contents/Developer/Library/
PrivateFrameworks to Contents/SharedFrameworks, breaking the @rpath load
of the helper binary.

CoreSimulator/SimulatorKit are only used via the Objective-C runtime, so
stop linking them in Package.swift and dlopen them from the active Xcode
instead (SimFrameworks.load), trying both the new SharedFrameworks and the
legacy PrivateFrameworks locations. This removes the version-specific
@rpath dependency entirely, so the DYLD_FRAMEWORK_PATH injection in the
launcher is no longer needed.

Also fix build.sh to resolve the product dir via `swift build
--show-bin-path` (Xcode 27 emits to .build/out/Products/Release).
