+++ packages/serve-sim/src/index.ts
@@ -202,27 +202,6 @@ function findHelperBinary(): string {
   return extracted;
 }
 
-/**
- * Env to spawn the Swift helper with. The helper links SimulatorKit/CoreSimulator
- * via `@rpath`, but the rpath baked in at build time points at whatever Xcode
- * lived on the build machine (e.g. `/Applications/Xcode_16.4.app/...`). On any
- * machine with Xcode installed at a different path that lookup fails with
- * `dyld: Library not loaded: @rpath/SimulatorKit.framework`. Inject the user's
- * actual Xcode PrivateFrameworks dir so dyld can resolve it regardless.
- */
-function helperSpawnEnv(): NodeJS.ProcessEnv {
-  let dev: string | null = null;
-  try {
-    dev = execSync("xcode-select -p", { encoding: "utf-8", stdio: ["ignore", "pipe", "ignore"] }).trim();
-  } catch {}
-  if (!dev) return process.env;
-  const fw = `${dev}/Library/PrivateFrameworks`;
-  return {
-    ...process.env,
-    DYLD_FRAMEWORK_PATH: process.env.DYLD_FRAMEWORK_PATH ? `${fw}:${process.env.DYLD_FRAMEWORK_PATH}` : fw,
-  };
-}
-
 // ─── Device helpers ───
 
 /**
@@ -465,7 +444,6 @@ async function spawnHelperDetached(opts: SpawnHelperOptions): Promise<{
   const child = nodeSpawn(helperPath, [udid, "--port", String(port)], {
     detached: true,
     stdio: ["ignore", logFd, logFd],
-    env: helperSpawnEnv(),
   });
   child.unref();
   closeSync(logFd);
@@ -498,7 +476,6 @@ async function spawnHelperAttached(opts: SpawnHelperOptions): Promise<{
   const child = nodeSpawn(helperPath, [udid, "--port", String(port)], {
     detached: false,
     stdio: ["ignore", logFd, logFd],
-    env: helperSpawnEnv(),
   });
   closeSync(logFd);
